pyproject.toml
[tool.poetry]
name = "User"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.2"
webdriver-manager = "^4.0.1"
undetected-chromedriver = "^3.5.5"
pyotp = "^2.9.0"
pycryptodome = "^3.20.0"
thefirstock = "^4.1.0"
kiteconnect = "^5.0.1"
pydantic = "^2.7.1"
pya3 = "^1.0.29"
firebase-admin = "^6.5.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pendulum = "^3.0.0"
pandas = "^1.5.3"
reportlab = "^3.6.0"
fastapi = "^0.99.0"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trademan = "User.main:main_api"
celery_worker = "User.Executor.Scripts.CeleryScripts.trademan_celery_app:start_worker"
celery_beat = "User.Executor.Scripts.CeleryScripts.trademan_celery_app:start_beat"
